requirements.txt
Flask
pymongo
requests 
python-dotenv